# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": ""
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/": {
      "get": {
        "tags": [
          "Root"
        ],
        "description": "End-point root da aplicação",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/produtos": {
      "post": {
        "tags": [
          "Produtos"
        ],
        "description": "Incluir um produto",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Listar um produto",
        "parameters": [
          {
            "name": "nome",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/produtos/{codigo}": {
      "get": {
        "tags": [
          "Produtos"
        ],
        "description": "Buscar um produto",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Produtos"
        ],
        "description": "Alterar um produto",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Produtos"
        ],
        "description": "Excluir um produto",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}